requirements.txt
colorama==0.3.7
colored==1.3.1
cprint==1.1
decorator==4.0.10
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
green==2.5.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
mock==2.0.0
pathlib2==2.1.0
pbr==1.10.0
pickleshare==0.7.4
prompt==0.4.0
prompt-toolkit==1.0.8
pyfiglet==0.7.5
Pygments==2.1.3
python-firebase==1.2
python-termstyle==0.1.10
requests==1.1.0
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
termcolor==1.1.0
traitlets==4.3.1
Unidecode==0.4.19
wcwidth==0.1.7
win-unicode-console==0.5
